scripts/mask2former/2nd-refine/exp01-bpr/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# exp01-bpr requirements
# exp04-GC/requirements.txt と同一。追加ライブラリなし。
# ─────────────────────────────────────────────────────────────────────────────

# PyTorch (CUDA 12.4 ビルド)
# インストール時は pip コマンドを分けて実行すること:
#   pip install torch==2.5.1 torchvision==0.20.1 --index-url https://download.pytorch.org/whl/cu124
torch==2.5.1
torchvision==0.20.1

# HuggingFace (Stage1 Mask2Former ロード用)
transformers==4.41.2
accelerate==0.30.0

# 画像処理・データ拡張
Pillow==10.3.0
albumentations>=2.0.0
albucore>=0.1.0
opencv-python-headless==4.9.0.80

# 数値計算・可視化
numpy==1.26.4
matplotlib==3.9.0

# ユーティリティ
tqdm==4.66.4
